refactor: Rearrange the project repository class and change its name to ensure better code readability and accessibility.
Change RemindersLocalRepository to RemindersRepository.
