+++ app/src/main/java/com/udacity/project4/data/repository/RemindersRepository.kt
@@ -1,4 +1,4 @@
-package com.udacity.project4.data.local
+package com.udacity.project4.data.repository
 
 import android.location.Location
 import com.google.android.gms.location.FusedLocationProviderClient
@@ -7,13 +7,13 @@ import com.udacity.project4.data.MyApp
 import com.udacity.project4.data.dto.ReminderDataSource
 import com.udacity.project4.data.dto.ReminderDTO
 import com.udacity.project4.data.dto.Result
+import com.udacity.project4.data.local.RemindersDao
 import com.udacity.project4.utils.AppSharedMethods
 import com.udacity.project4.utils.wrapEspressoIdlingResource
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ensureActive
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.tasks.await
@@ -27,7 +27,7 @@ import timber.log.Timber
  * @param remindersDao the dao that does the Room db operations
  * @param ioDispatcher a coroutine dispatcher to offload the blocking IO tasks
  */
-class RemindersLocalRepository(
+class RemindersRepository(
     private val remindersDao: RemindersDao,
     private val ioDispatcher: CoroutineDispatcher = Dispatchers.IO,
     private val fusedLocationProviderClient: FusedLocationProviderClient
